Initial implementation of models and persistence for managed channels (WIP)
This establishes most of the boilerplate necessary to begin having channel specific and channel+user specific settings. It is completely untested, unfinished, and nothing uses it yet, but is enough that it seems worthy of committing.

I will also mention that this doesn't feel very DRY, but is in the spirit of keeping things straightforward and clean for now.