New persistence layer that utilizes mybatis-ds-generator (WIP)
** Database schema changes **

This uses mybatis-ds-generator to generate improved model and mapper classes that use MyBatis Dynamic SQL to express more complex queries and relationships in pure java.

WIP - This implements most basic functionality that existed before but does not yet have an interface for managed channels. It also depends on an external library ds-generator-core-1.2.0.jar to be in the lib directory.